- #310 - fixed casting error that was causing local chatroom creation to crash - changed Chatroom userCount var to int on server

+++ java/src/com/chat/Chatroom.java
@@ -28,7 +28,7 @@ public class Chatroom implements Serializable {
     private final long radius;
     private final long creationTime;
     private short chatActivity;
-    private short userCount;
+    private int userCount;
     private final boolean isPrivate;
 
     // back-reference for easy access
